Update core
- linux: fix overclocking (issue #872)
- linux: add overclocking through "screen", if ocdaemon is not started/used

+++ OCDaemon.psm1
@@ -37,12 +37,12 @@ param(
     [Parameter(Mandatory = $False)]
     [Switch]$Quiet = $false
 )
-    if (-not $Cmd -and (-not $FilePath -or -not (Test-Path $FilePath))) {return}
+    if (-not ($Cmd -or $Global:GlobalOCD.Count) -and (-not $FilePath -or -not (Test-Path $FilePath))) {return}
 
     if (-not $Name) {$Name = (Get-ChildItem $FilePath -ErrorAction Ignore).BaseName}
 
     "$(Get-Date)" | Out-File "/opt/rainbowminer/ocdcmd/$Name.lock" -ErrorAction Ignore -Force
-    if ($Cmd) {
+    if ($Cmd -or $Global:GlobalOCD.Count) {
         Set-BashFile -FilePath "/opt/rainbowminer/ocdcmd/$Name.sh" -Cmd $Cmd
     } else {
         Copy-Item $FilePath "/opt/rainbowminer/ocdcmd/$Name.sh" -ErrorAction Ignore
@@ -99,6 +99,10 @@ param(
     [Parameter(Mandatory = $False)]
     $Cmd,
     [Parameter(Mandatory = $False)]
+    $FilePath,
+    [Parameter(Mandatory = $False)]
+    $Miner,
+    [Parameter(Mandatory = $False)]
     [Switch]$Quiet
 )
 
@@ -111,11 +115,50 @@ param(
     if (-not (Test-Path Variable:Global:GlobalOCD)) {[System.Collections.ArrayList]$Global:GlobalOCD = @()}
 
     if ($Cmd -or $Global:GlobalOCD.Count) {
-        $tmpfn = "$($Session.OCDaemonPrefix).$($Session.OCDaemonCount)"
-        Invoke-OCDaemonWithName -Name $tmpfn -Cmd $Cmd
-        if (Test-Path "/opt/rainbowminer/ocdcmd/$tmpfn.sh") {
-            Write-Log -Level Warn "OCDaemon failed. Please run `"./install.sh`" at the command line"
+        if (Test-OCDaemon) {
+            $tmpfn = "$($Session.OCDaemonPrefix).$($Session.OCDaemonCount)"
+            Invoke-OCDaemonWithName -Name $tmpfn -Cmd $Cmd
+            if (Test-Path "/opt/rainbowminer/ocdcmd/$tmpfn.sh") {
+                Write-Log -Level Warn "OCDaemon failed. Please run `"./install.sh`" at the command line"
+            }
+            $Session.OCDaemonCount++
+        } else {
+            $IsTemporaryPath = $false
+            if ($Miner) {
+                $ScreenName = ("$($Miner.DeviceName -join "_")" -replace "[^A-Z0-9_-]").ToLower()
+                Invoke-Exe "screen" -ArgumentList "-ls" -ExpandLines | Where-Object {$_ -match "(\d+)\.(oc_[a-z0-9_-]+)"} | Foreach-Object {
+                    $Name = $Matches[2]
+                    if ($Name -match "($(($ScreenName -split "_") -join "|"))") {
+                        Invoke-Exe "screen" -ArgumentList "-S $Name -X stuff `^C" > $null
+                        Start-Sleep -Milliseconds 250
+                        Invoke-Exe "screen" -ArgumentList "-S $Name -X quit" > $null
+                        Start-Sleep -Milliseconds 250
+                    }
+                }
+                $ScreenName = "oc_$ScreenName"
+                $FilePath = Join-Path (Split-Path $Miner.Path) "$($Miner.BaseAlgorithm -join "_")_$($ScreenName).sh"
+            } else {
+                if ($FilePath -eq "") {$FilePath = ".\Cache\tmp_$([System.IO.Path]::GetRandomFileName() -replace "\..+$").sh";$IsTemporaryPath = $true}
+                elseif ($FilePath -notmatch "\.sh$") {$FilePath = "$FilePath.sh"}
+                $ScreenName = "$(Split-Path $FilePath -Leaf)" -replace "\.sh$"
+                Invoke-Exe "screen" -ArgumentList "-ls" -ExpandLines | Where-Object {$_ -match "\d+\.$ScreenName\s+"} | Foreach-Object {
+                    Invoke-Exe "screen" -ArgumentList "-S $ScreenName -X stuff `^C" > $null
+                    Start-Sleep -Milliseconds 250
+                    Invoke-Exe "screen" -ArgumentList "-S $ScreenName -X quit" > $null
+                    Start-Sleep -Milliseconds 250
+                }
+            }
+
+            Set-BashFile -FilePath $FilePath -Cmd $Cmd
+
+            if (Test-Path $FilePath) {
+                (Start-Process "chmod" -ArgumentList "+x $FilePath" -PassThru).WaitForExit() > $null
+                Invoke-Exe "screen" -ArgumentList "-S $ScreenName -d -m" > $null
+                Start-Sleep 1
+                Invoke-Exe "screen" -ArgumentList "-S $ScreenName -X stuff $FilePath`n" > $null
+                Start-Sleep 1
+                if ($IsTemporaryPath) {Remove-Item $FilePath -Force -ErrorAction Ignore}
+            }
         }
-        $Session.OCDaemonCount++
     }
 }
